+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Areas only create, edit, delete by admin
 * User email no longer editable by owner
 * Fix extraData returning for Article by adding includeSite to the scope
+* Fix users can create/update/delete users with a higher role
 
 ## Unreleased
 * Statuscode for error 'Already voted' was 500 and is now 403